move download center def/links to one file
